customizing plugins generator

+++ config/actions/base/init.js
@@ -1,5 +1,5 @@
-var spawn = require('child_process').spawn;
 var path = require('path');
+var fs = require('fs');
 
 var prompts = require('./prompts');
 var myUtils = require('../../../util');
@@ -8,22 +8,35 @@ module.exports = function () {
 
     var cb = this.async();
 
+    var setConfig = function() {
+
+        /* and put better one inside `www` directory */
+        this.template('_config.xml', path.join('www', 'config.xml'));
+
+        cb();
+    };
+
     myUtils.success('Set config options:');
 
-    this.prompt(prompts.cordova, function (props) {
+    this.prompt(prompts, function (props) {
 
-        /* remove Cordova config */
-        fs.unlink('config.xml', function(err){
-            if(err) {
-                error('`rm config.xml` failed with code ' + errr.errno + '. Try to run it later.', 'red');
-            }
-            else {
-                /* and put better one inside `www` directory */
-                this.template('_config.xml', path.join('www', 'config.xml'));
+        this.preferences = props.preferences;
 
-                cb();
-            }
-        });
+        if(fs.existsSync('config.xml')) {
+
+            /* remove Cordova config */
+            fs.unlink('config.xml', function(err){
+                if(err) {
+                    error('`rm config.xml` failed with code ' + errr.errno + '. Try to run it later.', 'red');
+                }
+                else {
+                    setConfig.call(this);
+                }
+            }.bind(this));
+        }
+        else {
+            setConfig.call(this);
+        }
 
     }.bind(this));
 };
